chore: Rename s3 upload util file
