+++ app/util/__init__.py
@@ -1,1 +1,1 @@
-from util import *
+from asset_storage import *
